:truck: rename: footer, header 경로 변경
