+++ src/shared/components/Footer.tsx
